refactor: reorganize systems into folder structure
- Move all system files into dedicated folders (System/index.js pattern)
- Update all relative import paths to account for new directory depth
- Improve codebase organization and maintainability
