+++ src/js/systems/Camera/index.js
@@ -1,0 +1,1 @@
+export { CameraSystem } from './System.js';
