+++ src/js/systems/Connection/index.js
@@ -1,0 +1,1 @@
+export { ConnectionSystem } from './System.js';
