+++ src/js/systems/FPSController/index.js
@@ -1,0 +1,1 @@
+export { FPSControllerSystem } from './System.js';
