+++ src/js/systems/Level/index.js
@@ -1,0 +1,1 @@
+export { LevelLoader } from './Loader.js';
