+++ src/js/systems/Patrol/index.js
@@ -1,0 +1,1 @@
+export { PatrolSystem } from './System.js';
