+++ src/js/systems/Persistence/index.js
@@ -1,0 +1,1 @@
+export { PersistenceSystem } from './System.js';
