+++ src/js/systems/PlayerMovement/index.js
@@ -1,0 +1,1 @@
+export { PlayerMovementSystem } from './System.js';
