+++ src/js/systems/Render/index.js
@@ -1,0 +1,1 @@
+export { RenderSystem } from './System.js';
