+++ src/js/systems/ThreeRender/index.js
@@ -1,0 +1,1 @@
+export { ThreeRenderSystem } from './System.js';
