+++ src/js/systems/VoxelIndicatorRender/index.js
@@ -1,0 +1,1 @@
+export { VoxelIndicatorRenderSystem } from './System.js';
