+++ src/js/systems/index.js
@@ -1,13 +1,13 @@
-export { RenderSystem } from './RenderSystem.js';
-export { InputSystem } from './InputSystem.js';
-export { ThreeRenderSystem } from './ThreeRenderSystem.js';
-export { LevelLoader } from './LevelLoader.js';
-export { AgentSystem } from './AgentSystem.js';
-export { CameraSystem } from './CameraSystem.js';
-export { PlayerMovementSystem } from './PlayerMovementSystem.js';
-export { FPSControllerSystem } from './FPSControllerSystem.js';
-export { PatrolSystem } from './PatrolSystem.js';
-export { ConnectionSystem } from './ConnectionSystem.js';
-export { SessionSystem } from './SessionSystem.js';
-export { PersistenceSystem } from './PersistenceSystem.js';
-export { VoxelIndicatorRenderSystem } from './VoxelIndicatorRenderSystem.js';
+export { RenderSystem } from './Render/index.js';
+export { InputSystem } from './Input/index.js';
+export { ThreeRenderSystem } from './ThreeRender/index.js';
+export { LevelLoader } from './Level/index.js';
+export { AgentSystem } from './Agent/index.js';
+export { CameraSystem } from './Camera/index.js';
+export { PlayerMovementSystem } from './PlayerMovement/index.js';
+export { FPSControllerSystem } from './FPSController/index.js';
+export { PatrolSystem } from './Patrol/index.js';
+export { ConnectionSystem } from './Connection/index.js';
+export { SessionSystem } from './Session/index.js';
+export { PersistenceSystem } from './Persistence/index.js';
+export { VoxelIndicatorRenderSystem } from './VoxelIndicatorRender/index.js';
